Applied feedback comments
